stocksite/requirements.txt
asgiref==3.3.1
Django>=3.0,<4.0
djongo>=1.3
pymongo==3.11.3
pytz==2021.1
sqlparse==0.2.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
django-rest-auth==0.9.5
django-filter==2.4.0
django-redis==4.12.1
gunicorn==20.0.4
requests==2.25.1
django-debug-toolbar==3.2
gevent==21.1.2
